+++ src/sql/addresses.sql
@@ -1,11 +1,11 @@
 DROP TABLE IF EXISTS addresses;
 CREATE TABLE IF NOT EXISTS addresses (
-    action_index   INTEGER UNSIGNED NOT NULL, -- Unique action index
-    source_id      INTEGER UNSIGNED,          -- id of record in index_addresses table
-    fee_preference INTEGER UNSIGNED,
-    require_memo   INTEGER UNSIGNED,
-    memo_id        INTEGER UNSIGNED,          -- id of record in index_memos table 
-    status_id      INTEGER UNSIGNED           -- id of record in index_statuses table
+    action_index   BIGINT UNSIGNED NOT NULL, -- Unique action index
+    source_id      BIGINT UNSIGNED,          -- id of record in index_addresses table
+    fee_preference BIGINT UNSIGNED,
+    require_memo   BIGINT UNSIGNED,
+    memo_id        BIGINT UNSIGNED,          -- id of record in index_memos table 
+    status_id      BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index   ON addresses (action_index);
