+++ src/sql/mints.sql
@@ -1,12 +1,12 @@
 DROP TABLE IF EXISTS mints;
 CREATE TABLE mints (
-    action_index   INTEGER UNSIGNED NOT NULL, -- Unique action index
-    tick_id        INTEGER UNSIGNED,          -- id of record in index_ticks table
+    action_index   BIGINT UNSIGNED NOT NULL, -- Unique action index
+    tick_id        BIGINT UNSIGNED,          -- id of record in index_ticks table
     amount         VARCHAR(250),              -- Amount of token to mint
-    source_id      INTEGER UNSIGNED,          -- id of record in index_addresses table (address that did MINT)
-    destination_id INTEGER UNSIGNED,          -- id of record in index_addresses table (optional, mint and transfer)
-    memo_id        INTEGER UNSIGNED,          -- id of record in index_memos table 
-    status_id      INTEGER UNSIGNED           -- id of record in index_statuses table
+    source_id      BIGINT UNSIGNED,          -- id of record in index_addresses table (address that did MINT)
+    destination_id BIGINT UNSIGNED,          -- id of record in index_addresses table (optional, mint and transfer)
+    memo_id        BIGINT UNSIGNED,          -- id of record in index_memos table 
+    status_id      BIGINT UNSIGNED           -- id of record in index_statuses table
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE UNIQUE INDEX action_index   ON mints (action_index);
